Add ExperimentalPreviewHybridCommonAPI and gate/walk back Android leaks
Introduces a PHC-owned opt-in annotation for bridged native
experimental/internal APIs. Ad-tracking functions move to
experimentalCommon.kt behind the new annotation; other native
experimental/internal usages get a `phc:stable-bridge` acknowledgment
where PHC deliberately treats them as stable, established bridge
surface. Stale opt-ins (already stabilized upstream) are removed.
