+++ android/hybridcommon/config/detekt/detekt-baseline.xml
@@ -4,6 +4,7 @@
   <CurrentIssues>
     <ID>EmptyFunctionBlock:CommonKtTests.kt$CommonKtTests.&lt;no name provided&gt;${}</ID>
     <ID>Filename:common.kt$com.revenuecat.purchases.hybridcommon.common.kt</ID>
+    <ID>Filename:experimentalCommon.kt$com.revenuecat.purchases.hybridcommon.experimentalCommon.kt</ID>
     <ID>Filename:mappersHelpers.kt$com.revenuecat.purchases.hybridcommon.mappers.mappersHelpers.kt</ID>
     <ID>Filename:subscriberAttributes.kt$com.revenuecat.purchases.hybridcommon.subscriberAttributes.kt</ID>
     <ID>LargeClass:CommonKtTests.kt$CommonKtTests</ID>
